Fix "integrate" typo.
